probe class
