+++ package.json
@@ -6,10 +6,14 @@
 		"test:coverage:ci": "node node_modules/.bin/jest --coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
 	},
 	"devDependencies": {
+		"@babel/core": "^7.13.15",
+		"@babel/preset-env": "^7.13.15",
+		"@babel/preset-typescript": "^7.13.0",
 		"@types/cors": "^2.8.10",
 		"@types/express": "^4.17.11",
-		"@types/jest": "26.0.20",
+		"@types/jest": "^26.0.20",
 		"@types/node": "^14.14.37",
+		"babel-jest": "^26.6.3",
 		"esbuild": "0.9.2",
 		"husky": "^4.3.8",
 		"jest": "26.6.3",
