+++ probe/src/__tests__/probe._addSelector.test.js
@@ -1,4 +1,4 @@
-var Probe = require("../probe");
+var Probe = require("../probe").Probe;
 
 describe("_addSelector", () => {
 	test("basic test", () => {
