+++ server/handlers/probe-report.ts
@@ -1,0 +1,7 @@
+import type express from "express";
+
+export async function probeReport(req: express.Request, res: express.Response) {
+	console.info("API Hit");
+	// createJob("probe_report", req.body);
+	res.send("ack");
+}
